chore: MemoriesService 파일 분리
