+++ pochak/pochak/Network/Memories/MemoriesAPI.swift
@@ -41,87 +41,3 @@ extension MemoriesAPI: BaseAPI {
         }
     }
 }
-
-struct MemoriesService {
-    static let networkService = NetworkService.shared
-    
-    static func getMemorySummary(
-        userId: String,
-        completion: @escaping ( Result<CommonResponse<MemorySummary>, NetworkError>
-        ) -> Void
-    ) {
-        
-        networkService.requestWithCommonResponse(MemoriesAPI.getMemoriesSummary(userId),
-                                                 responseType: CommonResponse<MemorySummary>.self
-        ) { response in
-            switch response {
-            case .success(let data):
-                completion(.success(data))
-            case .failure(let error):
-                print("=== getMemorySummary service error ===")
-                print(error.localizedDescription)
-                completion(.failure(error))
-            }
-        }
-    }
-    
-    static func getMemoriesPochak(
-        userId: String,
-        completion: @escaping ( Result<CommonResponse<MemoryList>, NetworkError>
-        ) -> Void
-    ) {
-        
-        networkService.requestWithCommonResponse(MemoriesAPI.getPochakMemoryList(userId),
-                                                 responseType: CommonResponse<MemoryList>.self
-        ) { response in
-            switch response {
-            case .success(let data):
-                completion(.success(data))
-            case .failure(let error):
-                print("=== getMemoriesPochak service error ===")
-                print(error.localizedDescription)
-                completion(.failure(error))
-            }
-        }
-    }
-    
-    static func getMemoriesPochaked(
-        userId: String,
-        completion: @escaping ( Result<CommonResponse<MemoryList>, NetworkError>
-        ) -> Void
-    ) {
-        
-        networkService.requestWithCommonResponse(MemoriesAPI.getPochakedMemoryList(userId),
-                                                 responseType: CommonResponse<MemoryList>.self
-        ) { response in
-            switch response {
-            case .success(let data):
-                completion(.success(data))
-            case .failure(let error):
-                print("=== getMemoriesPochaked service error ===")
-                print(error.localizedDescription)
-                completion(.failure(error))
-            }
-        }
-    }
-    
-    static func getMemoriesBonded(
-        userId: String,
-        completion: @escaping ( Result<CommonResponse<MemoryList>, NetworkError>
-        ) -> Void
-    ) {
-        
-        networkService.requestWithCommonResponse(MemoriesAPI.getBondedMemoryList(userId),
-                                                 responseType: CommonResponse<MemoryList>.self
-        ) { response in
-            switch response {
-            case .success(let data):
-                completion(.success(data))
-            case .failure(let error):
-                print("=== getMemoriesBonded service error ===")
-                print(error.localizedDescription)
-                completion(.failure(error))
-            }
-        }
-    }
-}
